Add the ability to build with the bundled EVDI
The latest DisplayLink driver, 5.4, has included a version of EVDI that
works with kernels 5.11 and 5.12. The version that is currently released
on the EVDI github repo does not work with these kernels, even though
they have the same release version.

We can now issue a `make` to use the bundled version. If we want to
build the latest release from github, we can issue a `make unbundled` to
build with what is released in the future on github.

This patch also fixes the globals that were recently added to only be
set if we have not defined these macros through the Makefile. Otherwise,
defining our variables using the Makefile will have no affect on the rpmbuild.