Update to DisplayLink 6.1 Release
There is a newer package published by DisplayLink. This change updates to that
new release.

A patch from the evdi repo is being applied to the bundled version of evdi.
Added that patch when using the official 1.14.7 release on GitHub.

Update the udev rules file to match was is being installed by the official
DisplayLink driver package.

Signed-off-by: Michael L. Young <elgueromexicano@gmail.com>